requirements.txt
anonfiles-py==0.1.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
fastapi==0.61.2
gunicorn==20.0.4
h11==0.11.0
httpcore==0.12.2
httptools==0.1.1
httpx==0.16.1
idna==2.10
img2pdf==0.4.0
importlib-metadata==3.10.0
iniconfig==1.1.1
Jinja2==2.11.2
lxml==4.6.3
MarkupSafe==1.1.1
packaging==20.9
pikepdf==2.1.2
Pillow==8.0.1
pluggy==0.13.1
py==1.10.0
pydantic==1.7.2
pymongo==3.11.2
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.0.1
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.12.3
uvloop==0.14.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.2.0
